pyproject.toml
[project]
name = "temporary-google-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
